# packages/contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/google": {
      "post": {
        "operationId": "AuthController_google",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleLoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicUserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/events": {
      "post": {
        "operationId": "EventsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "events"
        ]
      }
    },
    "/api/events/mine": {
      "get": {
        "operationId": "EventsController_listMine",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MyEventDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "events"
        ]
      }
    },
    "/api/events/{slug}": {
      "get": {
        "operationId": "EventsController_getBySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventDetailDto"
                }
              }
            }
          }
        },
        "tags": [
          "events"
        ]
      }
    },
    "/api/events/{id}": {
      "patch": {
        "operationId": "EventsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "events"
        ]
      }
    },
    "/api/events/{id}/publish": {
      "post": {
        "operationId": "EventsController_publish",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "events"
        ]
      }
    },
    "/api/events/{id}/ticket-types": {
      "post": {
        "operationId": "EventsController_addTicketType",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTicketTypeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketTypePublicDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "events"
        ]
      }
    },
    "/api/events/{id}/ticket-types/{ticketTypeId}": {
      "patch": {
        "operationId": "EventsController_updateTicketType",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ticketTypeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTicketTypeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketTypePublicDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "events"
        ]
      }
    },
    "/api/events/{id}/checkin": {
      "post": {
        "operationId": "CheckinController_scan",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckinDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckinResultDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "checkin"
        ]
      }
    },
    "/api/events/{id}/attendees.csv": {
      "get": {
        "operationId": "CheckinController_csv",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "checkin"
        ]
      }
    },
    "/api/events/{eventId}/holds": {
      "post": {
        "operationId": "HoldsController_acquire",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-hold-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-admission-token",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcquireHoldDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HoldDto"
                }
              }
            }
          }
        },
        "tags": [
          "holds"
        ]
      }
    },
    "/api/events/{eventId}/holds/mine": {
      "get": {
        "operationId": "HoldsController_listMine",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-hold-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HoldDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "holds"
        ]
      }
    },
    "/api/events/{eventId}/holds/{seatId}": {
      "delete": {
        "operationId": "HoldsController_release",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "seatId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-hold-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "holds"
        ]
      }
    },
...
  },
  "info": {
    "title": "OpenSeat API",
    "description": "REST API for the OpenSeat ticketing platform",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "ada@example.com"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 72
          },
          "displayName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80,
            "example": "Ada Lovelace"
          }
        },
        "required": [
          "email",
          "password",
          "displayName"
        ]
      },
      "PublicUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "isDemo": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "email",
          "displayName",
          "isDemo"
        ]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/PublicUserDto"
          },
          "accessToken": {
            "type": "string"
          }
        },
        "required": [
          "user",
          "accessToken"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "ada@example.com"
          },
          "password": {
            "type": "string",
            "maxLength": 72
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "GoogleLoginDto": {
        "type": "object",
        "properties": {
          "credential": {
            "type": "string",
            "description": "Google Identity Services ID token (JWT)"
          }
        },
        "required": [
          "credential"
        ]
      },
...
      "CreateTicketTypeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80,
            "example": "General admission"
          },
          "quantity": {
            "type": "number",
            "minimum": 1,
            "maximum": 100000
          },
          "priceSatang": {
            "type": "number",
            "minimum": 0,
            "maximum": 100000000,
            "description": "Price in satang; 0 = free"
          },
          "maxPerOrder": {
            "type": "number",
            "minimum": 1,
            "maximum": 20,
            "default": 10
          }
        },
        "required": [
          "name",
          "quantity",
          "priceSatang"
        ]
      },
      "CreateEventDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 3,
            "maxLength": 120,
            "example": "Bangkok Indie Fest"
          },
          "description": {
            "type": "string",
            "maxLength": 5000,
            "default": ""
          },
          "venueName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 160,
            "example": "Voice Space, Bangkok"
          },
          "startsAt": {
            "type": "string",
            "example": "2026-09-01T12:00:00.000Z"
          },
          "endsAt": {
            "type": "string",
            "example": "2026-09-01T16:00:00.000Z"
          },
          "ticketTypes": {
            "minItems": 1,
            "maxItems": 10,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateTicketTypeDto"
            }
          }
        },
        "required": [
          "title",
          "venueName",
          "startsAt",
          "ticketTypes"
        ]
      },
      "EventOrganizerDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "displayName"
        ]
      },
      "TicketTypePublicDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "kind": {
            "enum": [
              "ga",
              "seated"
            ],
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "priceSatang": {
            "type": "number"
          },
          "quantity": {
            "type": "number"
          },
          "remaining": {
            "type": "number"
          },
          "maxPerOrder": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "kind",
          "name",
          "priceSatang",
          "quantity",
          "remaining",
          "maxPerOrder"
        ]
      },
      "EventSeatMapRefDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "required": [
          "id"
        ]
      },
      "EventDetailDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string"
          },
          "endsAt": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "enum": [
              "draft",
              "published"
            ],
            "type": "string"
          },
          "coverImageUrl": {
            "type": "string",
            "nullable": true
          },
          "isDemo": {
            "type": "boolean"
          },
          "dropMode": {
            "type": "boolean"
          },
          "saleOpensAt": {
            "type": "string",
            "nullable": true
          },
          "organizer": {
            "$ref": "#/components/schemas/EventOrganizerDto"
          },
          "ticketTypes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TicketTypePublicDto"
            }
          },
          "seatMap": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/EventSeatMapRefDto"
              }
            ]
          },
          "id": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "venueName": {
            "type": "string"
          },
          "startsAt": {
            "type": "string"
          }
        },
        "required": [
          "description",
          "endsAt",
          "status",
          "coverImageUrl",
          "isDemo",
          "dropMode",
          "saleOpensAt",
          "organizer",
          "ticketTypes",
          "seatMap",
          "id",
          "slug",
          "title",
          "venueName",
          "startsAt"
        ]
      },
      "MyEventDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string"
          },
          "endsAt": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "published"
            ]
          },
          "coverImageUrl": {
            "type": "string",
            "nullable": true
          },
          "isDemo": {
            "type": "boolean"
          },
          "dropMode": {
            "type": "boolean"
          },
          "saleOpensAt": {
            "type": "string",
            "nullable": true
          },
          "ticketTypes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TicketTypePublicDto"
            }
          },
          "seatMap": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/EventSeatMapRefDto"
              }
            ]
          },
          "ticketsIssued": {
            "type": "number"
          },
          "id": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "venueName": {
            "type": "string"
          },
          "startsAt": {
            "type": "string"
          }
        },
        "required": [
          "description",
          "endsAt",
          "status",
          "coverImageUrl",
          "isDemo",
          "dropMode",
          "saleOpensAt",
          "ticketTypes",
          "seatMap",
          "ticketsIssued",
          "id",
          "slug",
          "title",
          "venueName",
          "startsAt"
        ]
      },
      "UpdateEventDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 3,
            "maxLength": 120
          },
          "description": {
            "type": "string",
            "maxLength": 5000
          },
          "venueName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 160
          },
          "startsAt": {
            "type": "string"
          },
          "endsAt": {
            "type": "string"
          }
        }
      },
      "UpdateTicketTypeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80
          },
          "quantity": {
            "type": "number",
            "minimum": 1,
            "maximum": 100000
          },
          "maxPerOrder": {
            "type": "number",
            "minimum": 1,
            "maximum": 20
          }
        }
      },
      "CheckinDto": {
        "type": "object",
        "properties": {
          "qrToken": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "description": "The scanned ticket QR token"
          }
        },
        "required": [
          "qrToken"
        ]
      },
      "CheckinResultDto": {
        "type": "object",
        "properties": {
          "outcome": {
            "type": "string",
            "enum": [
              "checked_in",
              "already_checked_in"
            ]
          },
          "ticketId": {
            "type": "string"
          },
          "attendeeName": {
            "type": "string"
          },
          "ticketType": {
            "type": "string"
          },
          "seat": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string"
          },
          "checkedInAt": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "outcome",
          "ticketId",
          "attendeeName",
          "ticketType",
          "seat",
          "status",
          "checkedInAt"
        ]
      },
      "AcquireHoldDto": {
        "type": "object",
        "properties": {
          "seatId": {
            "type": "string"
          }
        },
        "required": [
          "seatId"
        ]
      },
      "HoldDto": {
        "type": "object",
        "properties": {
          "seatId": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string"
          }
        },
        "required": [
          "seatId",
          "expiresAt"
        ]
      },
...
    }
  }
}